helloWebApp/helloWebApp/src/main/java/com/joseph/gradle/UsernameServlet.java
package com.joseph.gradle;

import java.io.IOException;
// import jakarta.servlet.ServletConfig;
// import jakarta.servlet.ServletContext;
// import jakarta.servlet.ServletException;
// import jakarta.servlet.annotation.WebServlet;
// import jakarta.servlet.http.HttpServlet;
// import jakarta.servlet.http.HttpServletRequest;
// import jakarta.servlet.http.HttpServletResponse;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.annotation.WebServlet;

@WebServlet("/username")
public class UsernameServlet extends HttpServlet {
    protected void doPost(HttpServletRequest request, HttpServletResponse response)
        throws ServletException, IOException {
            String username = request.getParameter("username");
            request.setAttribute("helloMessage", "Hello " + username);
            request.getRequestDispatcher("response.jsp").forward(request, response);
     }
}